// src/app/models/app-user.ts
export class AppUser {
  public _id: string;
  public Username: string;
  public Email: string;
  public Password: string;
  public ConfirmPassword: string;
  public PhoneNumber: number;
  public Address: string;
  public State: string;
  public Country: string;
  public Role: string;
  public ImageUrl: string;

}

export interface AppUser {
  _id: string;
  Username: string;
  Email: string;
  Password: string;
  ConfirmPassword: string;
  PhoneNumber: number;
  Address: string;
  State: string;
  Country: string;
  Role: string;
  ImageUrl: string;

}